==> 2_Fetch_Ticker_Data.ipynb <==
import requests
import pandas as pd
import matplotlib.pyplot as plt

#!pip install alpha_vantage
from alpha_vantage.timeseries import TimeSeries
from alpha_vantage.techindicators import TechIndicators
from alpha_vantage.fundamentaldata import FundamentalData 
ts = TimeSeries(key = 'HKDDDB1WO21X9QR6', output_format = 'pandas', indexing_type = 'date')
ti = TechIndicators(key = 'HKDDDB1WO21X9QR6', output_format = 'pandas', indexing_type = 'date')
fd = FundamentalData(key = 'HKDDDB1WO21X9QR6', output_format = 'pandas', indexing_type = 'date')

from functools import reduce

ticker = 'ZM'

key = 'HKDDDB1WO21X9QR6'
base_url = "https://www.alphavantage.co/query/"

#build out a request
res = requests.get(url=base_url, params={
    "function":"EARNINGS",
    "symbol":f'{ticker}',
    'apikey':key})

res.status_code

EPS = res.json()
# _, header = EPS.json()
#df = pd.DataFrame.from_dict(EPS, orient= 'index')_

len(EPS['quarterlyEarnings'])

def quarterly_earnings_df(ticker):
    key = 'HKDDDB1WO21X9QR6'
    base_url = "https://www.alphavantage.co/query/"
    stock = ticker
    res = requests.get(url=base_url, params={
    "function":"EARNINGS",
    "symbol":f'{stock}',
    'apikey':key})
    EPS = res.json()
    
    #Quarterly
    for i in range(0,len(EPS['quarterlyEarnings'])-1):
        if i==0:
            df = pd.DataFrame.from_dict(EPS['quarterlyEarnings'][i], orient = 'index')
        else:
            df_1 = pd.DataFrame.from_dict(EPS['quarterlyEarnings'][i], orient = 'index')
            df = pd.merge(df, df_1, left_index = True, right_index = True) 
    df = df.transpose()
    df['fiscalDateEnding'] = pd.to_datetime(df['fiscalDateEnding'])
    df['reportedDate'] = pd.to_datetime(df['reportedDate'])
    df[['reported_Quarterly_EPS', 'estimated_Quarterly_EPS', 'surprise', 'surprisePercentage']] = df[['reportedEPS', 'estimatedEPS', 'surprise', 'surprisePercentage']].apply(pd.to_numeric, errors = 'coerce')
    df = df.drop(columns = ['estimatedEPS', 'reportedEPS'])
    df = df.set_index('fiscalDateEnding')
    return df

df = quarterly_earnings_df(ticker)

df.head()

reported_dict = pd.Series(df.reportedDate.values, index = df.index).to_dict()    

def annual_earnings_df(ticker):
    key = 'HKDDDB1WO21X9QR6'
    base_url = "https://www.alphavantage.co/query/"
    stock = ticker
    res = requests.get(url=base_url, params={
    "function":"EARNINGS",
    "symbol":f'{stock}',
    'apikey':key})
    EPS = res.json()
    #Annual
    for i in range(0,len(EPS['annualEarnings'])-1):
        if i==0:
            df = pd.DataFrame.from_dict(EPS['annualEarnings'][i], orient = 'index')
        else:
            df_1 = pd.DataFrame.from_dict(EPS['annualEarnings'][i], orient = 'index')
            df = pd.merge(df, df_1, left_index = True, right_index = True)
    df = df.transpose()
    df['fiscalDateEnding'] = pd.to_datetime(df['fiscalDateEnding'])
    df[['reported_Annual_EPS']] = df[['reportedEPS']].apply(pd.to_numeric, errors = 'coerce')
    df = df.drop(columns = 'reportedEPS')
    df = df.set_index('fiscalDateEnding')
    return df

df = annual_earnings_df(ticker)
#change index to the reported date for annual earnings 
#df.index = df.index.map(reported_dict)
df.head()

def all_earnings_df(ticker):
    key = 'HKDDDB1WO21X9QR6'
    base_url = "https://www.alphavantage.co/query/"
    stock = ticker
    res = requests.get(url=base_url, params={
    "function":"EARNINGS",
    "symbol":f'{stock}',
    'apikey':key})
    EPS = res.json()
    #Quarterly
    for i in range(0,len(EPS['quarterlyEarnings'])-1):
        if i==0:
            df = pd.DataFrame.from_dict(EPS['quarterlyEarnings'][i], orient = 'index')
        else:
            df_1 = pd.DataFrame.from_dict(EPS['quarterlyEarnings'][i], orient = 'index')
            df = pd.merge(df, df_1, left_index = True, right_index = True) 
    df = df.transpose()
    df['fiscalDateEnding'] = pd.to_datetime(df['fiscalDateEnding'])
    df['reportedDate'] = pd.to_datetime(df['reportedDate'])
    df[['reported_Quarterly_EPS', 'estimated_Quarterly_EPS', 'surprise', 'surprisePercentage']] = df[['reportedEPS', 'estimatedEPS', 'surprise', 'surprisePercentage']].apply(pd.to_numeric, errors = 'coerce')
    df = df.drop(columns = ['estimatedEPS', 'reportedEPS'])
    df = df.set_index('fiscalDateEnding')
    
    #create dictionary with fiscaldateending (key) and reporteddate (value) to map onto annual 
    reported_dict = pd.Series(df.reportedDate.values, index = df.index).to_dict()    
    
    #change index of quarterly to reported
    df = df.set_index('reportedDate')
    
    #Annual 
    res = requests.get(url=base_url, params={
    "function":"EARNINGS",
    "symbol":f'{stock}',
    'apikey':key})
    EPS = res.json()
    #Annual
    for i in range(0,len(EPS['annualEarnings'])-1):
        if i==0:
            dfa = pd.DataFrame.from_dict(EPS['annualEarnings'][i], orient = 'index')
        else:
            dfa_1 = pd.DataFrame.from_dict(EPS['annualEarnings'][i], orient = 'index')
            dfa = pd.merge(dfa, dfa_1, left_index = True, right_index = True)
    dfa = dfa.transpose()
    dfa['fiscalDateEnding'] = pd.to_datetime(dfa['fiscalDateEnding'])
    dfa[['reported_Annual_EPS']] = dfa[['reportedEPS']].apply(pd.to_numeric, errors = 'coerce')
    dfa = dfa.drop(columns = 'reportedEPS')
    dfa = dfa.set_index('fiscalDateEnding')
    
    #map new index values onto annual so that date is the reported date (more relevant to price action)
    dfa.index = dfa.index.map(reported_dict)

    #merge
    merge = pd.merge(df, dfa, how ='left', left_index = True, right_index = True)
    return merge

df = all_earnings_df(ticker)

df.head()

#Create funtion with other functions to get both annual and quarterly earnings into one DF

def get_all_earnings(ticker):
    dfq = quarterly_earnings_df(ticker)
    reported_dict = pd.Series(dfq.reportedDate.values, index = dfq.index).to_dict()    
    dfq = dfq.set_index('reportedDate')
    dfa = annual_earnings_df(ticker)
    dfa.index = dfa.index.map(reported_dict)
    merge = pd.merge(dfq, dfa, how ='left', left_index = True, right_index = True)
    return merge

earnings = get_all_earnings(ticker)

earnings.head(20)

earnings = earnings.sort_index()

#Export earnings data to data folder
earnings.to_csv(f'data/{ticker}_earnings.csv')

#### Get Daily Stock Data for Ticker along with Technical Indicators and IWO Index Values 

For Ticker:
- Open 
- High
- Low
- Close
- Volume

Techincal Indicators:
- 8 Day Exponential Moving Average
- 21 Day Exponential Moving Average
- 50 Day Simple Moving Average
- 200 Day Simple Moving Average

IWO Index:
- 5 Day Exponential Moving Average
- 10 Day Exponential Moving Average
- 150 Day Exponential Moving Average

# Get daily stock data
daily, meta_data = ts.get_daily(symbol = ticker, outputsize = 'full')
#UNCOMMENT TO MERGE EARNINGS DATA WITH PRICE ACTION ----- merge = pd.merge(daily, df, how ='outer', left_index = True, right_index = True)

#add IWO stock data for 5day EMA, 10day EMA and 150day EMA
data_5, meta_data = ti.get_ema(symbol = 'IWO', interval='daily', time_period = 5, series_type = 'close')
data_10, meta_data = ti.get_ema(symbol = 'IWO', interval='daily', time_period = 10, series_type = 'close')
data_150, meta_data = ti.get_ema(symbol = 'IWO', interval='daily', time_period = 150, series_type = 'close')
dfs = [data_5, data_10, data_150]
trend = reduce(lambda left, right: pd.merge(left, right, on='date', how = 'inner'), dfs)
trend.columns = ['IWO_5d_EMA', 'IWO_10d_EMA', 'IWO_150d_EMA']

#Inner merge so that we only get IWO data for when the ticker has price data as well (in case ticker recently public)
merge = pd.merge(daily, trend, how ='inner',left_index = True, right_index = True)
merge.head()

merge.tail()

# Create list of parameter dictionaries
def build_params(ticker):
    param_list = [{'EMA_8' : {
    "function":"EMA",
    "interval":"daily",
    "symbol":f'{ticker}',
    "time_period": "8",
    "series_type": "high",
    'apikey':key}}, 
 
    {'EMA_21':{
    "function":"EMA",
    "interval":"daily",
    "symbol":f'{ticker}',
    "time_period": "21",
    "series_type": "high",
    'apikey':key}}, 
                        
    {'SMA_50':{
    "function":"SMA",
    "interval":"daily",
    "symbol":f'{ticker}',
    "time_period": "50",
    "series_type": "high",
    'apikey':key}},
              
    {'SMA_200':{
    "function":"SMA",
    "interval":"daily",
    "symbol":f'{ticker}',
    "time_period": "200",
    "series_type": "high",
    'apikey':key}}
             
    ]
    return param_list

def get_technicals(stock):
    key = 'HKDDDB1WO21X9QR6'
    base_url = "https://www.alphavantage.co/query/"
    ticker = stock
    dfs = []
    cols = []
    param_list = build_params(ticker)
    for i in param_list:    
        for k, v in i.items():
            res = requests.get(url=base_url, params = v)
            res_dict = res.json()
            _, header = res.json()
            df = pd.DataFrame.from_dict(res_dict[header], orient= 'index')
            dfs.append(df)
            cols.append(k)
    df = pd.concat(dfs, axis = 1)
    df.columns = cols
    df = df.apply(pd.to_numeric, errors = 'coerce')
    return df
            

technicals = get_technicals(ticker)
stock_data = pd.merge(merge, technicals, how ='inner',left_index = True, right_index = True)

stock_data.tail()

stock_data.head()

#Sort index by datetime
stock_data = stock_data.sort_index()

#Export stock data to data folder
stock_data.to_csv(f'data/{ticker}_stock_data.csv')
