==> 6_CANSLIM.ipynb <==
import pandas as pd
import numpy as np

ticker = 'ZM'

#Read in earnings data for ticker as df
df = pd.read_csv(f'data/{ticker}_earnings.csv', index_col = 'reportedDate')
df.index = pd.to_datetime(df.index)

df

#Read in Additional CANSLIM inputs from Finviz notebook as df1
df1 = pd.read_csv(f'data/{ticker}_CANSLIM_data.csv', index_col = 'date')
df1.index = pd.to_datetime(df1.index)

df1

#Merge dataframes on datetime index
merge = pd.merge(df, df1, how ='outer',left_index = True, right_index = True)

merge

merge = merge.drop(columns = ['IWO_5d_EMA', 'IWO_10d_EMA', 'IWO_150d_EMA'])

stock_data = pd.read_csv(f'data/{ticker}_stock_data.csv', index_col = 'Unnamed: 0')

stock_data

merged = pd.merge(merge, stock_data, how ='outer',left_index = True, right_index = True)
merged.head()

merged.tail()

#Build fuction to output which CANSLIM criteria are met for ticker

def canslim(df):
    annual_earnings = []
    quarterly_earnings = []
    criteria = []
    keys = ['C', 'A', 'S', 'I', 'M']
    
    #C - earnings growth acclerating
    for i in df['reported_Quarterly_EPS']:
        if np.isnan(i) == False:
            quarterly_earnings.append(i)
    try:
        if quarterly_earnings[-1]> 1.25 * quarterly_earnings[-2] and quarterly_earnings[-2] > 1.25* quarterly_earnings[-3]:
            criteria.append(1)
        else: 
            criteria.append(0)
    except: 
        print("Limited Quarterly Earnings Data")
        if quarterly_earnings[-1]> 1.25 * quarterly_earnings[-2]:
            criteria.append(1)
        else: 
            criteria.append(0)               
    #A 
    annual_earnings = []
    for j in df['reported_Annual_EPS']:
        if np.isnan(j) == False:
            annual_earnings.append(j)
    try:
        if annual_earnings[-1]> 1.25 * annual_earnings[-2]: 
            criteria.append(1)
        else: 
            criteria.append(0)
    except: 
        print("Limited Annual Earnings Data")
        criteria.append(0)
    
    #N
    #Hold for future addition scraping most recent 10-K for instance of "new product" 
    
    #S - if mean price and volume in last 90 trading days is higher than the 90 before criteria met
    try:
        if df['4. close'][-90:].mean() > df['4. close'][-180:-90].mean() and df['5. volume'][-90:].mean() > df['5. volume'][-180:-90].mean():
            criteria.append(1)
        else:
            criteria.append(0)
    except:
        print("Limited Price and Volume Data")
        criteria.append(0)

    #L
    #Hold for future addition using industry group rank
    
    #I
    try:
        if df['Institutional_Ownership'][-1]>0.20 and df['Change_Instituional_Ownership_3months'][-1]>0:
            criteria.append(1)
        else:
            criteria.append(0)
    except:
        print('Limited Institutional Ownership Information')
        criteria.append(0)    
    #M
    if df['IWO_5d_EMA'][-1] > merged['IWO_10d_EMA'][-1] and merged['IWO_10d_EMA'][-1]> merged['IWO_150d_EMA'][-1]:
        criteria.append(1)
    else:
        criteria.append(0)
    canslim_dict = dict(zip(keys, criteria))
    return canslim_dict
   

canslim(merged)

df_dict = canslim(merged)

df = pd.DataFrame(df_dict, index = [ticker])

df
